+++ src/pytest_datadir/plugin.py
@@ -16,7 +16,17 @@ def _win32_longpath(path):
         # to a path string tells the Windows APIs to disable all string parsing
         # and to send the string that follows it straight to the file system".
         # (See https://docs.microsoft.com/pt-br/windows/desktop/FileIO/naming-a-file)
-        return "\\\\?\\" + os.path.normpath(path)
+        normalized = os.path.normpath(path)
+        if not normalized.startswith("\\\\?\\"):
+            is_unc = normalized.startswith("\\\\")
+            # see https://en.wikipedia.org/wiki/Path_(computing)#Universal_Naming_Convention # noqa: E501
+            if (
+                is_unc
+            ):  # then we need to insert an additional "UNC\" to the longpath prefix
+                normalized = normalized.replace("\\\\", "\\\\?\\UNC\\")
+            else:
+                normalized = "\\\\?\\" + normalized
+        return normalized
     else:
         return path